// src/core/domain/enums/FeaturedPostCategory.ts
enum FeaturedPostCategory {
  HEALTH = "HEALTH",
  EDUCATION = "EDUCATION",
  TECHNOLOGY = "TECHNOLOGY",
  ENVIRONMENT = "ENVIRONMENT",
  ENTREPRENEURSHIP = "ENTREPRENEURSHIP",
  SELF_DEVELOPMENT = "SELF_DEVELOPMENT",
  ARTS = "ARTS",
  SPORTS = "SPORTS",
  ENTERTAINMENT = "ENTERTAINMENT",
  DISABILITY = "DISABILITY",
  ECONOMY = "ECONOMY",
  LAW = "LAW",
}

export default FeaturedPostCategory